python/requirements.txt
# Core dependencies (pinned to installed versions)
Pillow==12.0.0
requests==2.32.5
orjson==3.11.5
PyPDF2==3.0.1
img2pdf==0.6.3